moved menu bar
